+++ src/components/compound/Board/types.ts
