+++ src/components/pages/Start/index.tsx
@@ -1,0 +1,9 @@
+import Component from "./component";
+import { useStartPageData } from "./data";
+
+const Start = () => {
+  const { mode, setMode } = useStartPageData();
+  return <Component mode={mode} setMode={setMode} />;
+};
+
+export default Start;
